requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy
stripe==14.4.0
python-multipart==0.0.22
jinja2
python-dotenv
authlib
httpx
itsdangerous
passlib[bcrypt]
bcrypt
